finalize incentive system

+++ database/migrations/2026_02_25_005501_alter_parties_table.php
@@ -4,20 +4,30 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration {
+return new class extends Migration
+{
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::table('parties', function (Blueprint $table) {
-            $table->json('role')->nullable()->after('type');
-            $table->integer('old_id')->nullable()->after('role');
-        });
-        Schema::table('matter_party', function (Blueprint $table) {
-            $table->string('role')->nullable()->after('parent_id');
-        });
-
+        if (Schema::hasTable('parties')) {
+            Schema::table('parties', function (Blueprint $table) {
+                if (! Schema::hasColumn('parties', 'role')) {
+                    $table->json('role')->nullable()->after('type');
+                }
+                if (! Schema::hasColumn('parties', 'old_id')) {
+                    $table->integer('old_id')->nullable()->after('role');
+                }
+            });
+        }
+        if (Schema::hasTable('matter_party')) {
+            Schema::table('matter_party', function (Blueprint $table) {
+                if (! Schema::hasColumn('matter_party', 'role')) {
+                    $table->string('role')->nullable()->after('parent_id');
+                }
+            });
+        }
     }
 
     /**
@@ -25,12 +35,16 @@ return new class extends Migration {
      */
     public function down(): void
     {
-        Schema::table('parties', function (Blueprint $table) {
-            $table->dropColumn('role');
-            $table->dropColumn('old_id');
-        });
-        Schema::table('matter_party', function (Blueprint $table) {
-            $table->dropColumn('role');
-        });
+        if (Schema::hasTable('parties')) {
+            Schema::table('parties', function (Blueprint $table) {
+                $table->dropColumn('role');
+                $table->dropColumn('old_id');
+            });
+        }
+        if (Schema::hasTable('matter_party')) {
+            Schema::table('matter_party', function (Blueprint $table) {
+                $table->dropColumn('role');
+            });
+        }
     }
 };
