+++ database/migrations/2026_03_11_015741_add_type_for_attachment_table.php
@@ -1,23 +1,36 @@
 <?php
 
+use App\Models\User;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration {
+return new class extends Migration
+{
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::table('attachments', function (Blueprint $table) {
-            $table->foreignIdFor(\App\Models\User::class)->after('matter_id')->constrained()->onDelete('cascade');
-            $table->string('path')->after('name')->nullable();
-            $table->string('type')->after('name')->nullable();
-            $table->renameColumn('mime', 'size');
-            $table->renameColumn('extention', 'extension');
-            $table->timestamps();
-        });
+        if (Schema::hasTable('attachments')) {
+            Schema::table('attachments', function (Blueprint $table) {
+                if (! Schema::hasColumn('attachments', 'user_id')) {
+                    $table->foreignIdFor(User::class)->after('matter_id')->constrained()->onDelete('cascade');
+                }
+                if (! Schema::hasColumn('attachments', 'path')) {
+                    $table->string('path')->after('name')->nullable();
+                }
+                if (! Schema::hasColumn('attachments', 'type')) {
+                    $table->string('type')->after('name')->nullable();
+                }
+                if (Schema::hasColumn('attachments', 'mime')) {
+                    $table->renameColumn('mime', 'size');
+                }
+                if (Schema::hasColumn('attachments', 'extention')) {
+                    $table->renameColumn('extention', 'extension');
+                }
+            });
+        }
     }
 
     /**
@@ -25,13 +38,24 @@ return new class extends Migration {
      */
     public function down(): void
     {
-        Schema::table('attachments', function (Blueprint $table) {
-            $table->dropColumn('type');
-            $table->dropColumn('path');
-            $table->dropConstrainedForeignId('user_id');
-            $table->renameColumn('size', 'mime');
-            $table->renameColumn('extension', 'extentions');
-            $table->dropTimestamps();
-        });
+        if (Schema::hasTable('attachments')) {
+            Schema::table('attachments', function (Blueprint $table) {
+                if (Schema::hasColumn('attachments', 'type')) {
+                    $table->dropColumn('type');
+                }
+                if (Schema::hasColumn('attachments', 'path')) {
+                    $table->dropColumn('path');
+                }
+                if (Schema::hasColumn('attachments', 'user_id')) {
+                    $table->dropConstrainedForeignId('user_id');
+                }
+                if (Schema::hasColumn('attachments', 'size')) {
+                    $table->renameColumn('size', 'mime');
+                }
+                if (Schema::hasColumn('attachments', 'extension')) {
+                    $table->renameColumn('extension', 'extentions');
+                }
+            });
+        }
     }
 };
